Application/Backend/requirements.txt
# Digital-Ḍād OCR — Backend Dependencies
# pip install -r requirements.txt

# Web
flask==3.0.3
flask-cors==4.0.1

# Vision / image processing
opencv-python==4.10.0.84
Pillow==10.4.0
numpy==1.26.4
scikit-image==0.24.0

# Deep learning
tensorflow==2.15.0
keras==2.15.0

# DocTr (geometric correction)
torch==2.3.1
torchvision==0.18.1
timm==1.0.9

# Google Cloud OCR (OCR.py)
google-cloud-vision==3.7.4

# Vertex AI / Gemini (Grade.py)
google-cloud-aiplatform==1.60.0

# Metrics
python-Levenshtein==0.25.1
